Change the name of all functions in the secp256k1 C source code to avoid linking issues
When linking the Rust and Go libraries together for the VM fuzzer, there was the
problem that both libraries export the same symbols; both embed the secp256k1
C library, and the linker complained that there were "multiple definitions".
By altering the function names in the Geth, the linking error is avoided
as there are no longer duplicate symbols.